+++ cli/tests/workers/permissions_data_local.ts.out
@@ -1,4 +1,4 @@
 error: Uncaught (in worker "") Requires read access to "[WILDCARD]local_file.ts", run again with the --allow-read flag
     at data:application/javascript;base64,[WILDCARD]:1:0
 error: Uncaught (in promise) Error: Unhandled error event reached main worker.
-    at Worker.#pollCtrl ([WILDCARD])
+    at Worker.#pollControl ([WILDCARD])
