+++ cli/tests/workers/permissions_data_remote.ts.out
@@ -1,4 +1,4 @@
 error: Uncaught (in worker "") Requires net access to "example.com", run again with the --allow-net flag
     at data:application/javascript;base64,aW1wb3J0ICJodHRwczovL2V4YW1wbGUuY29tL3NvbWUvZmlsZS50cyI7:1:0
 error: Uncaught (in promise) Error: Unhandled error event reached main worker.
-    at Worker.#pollCtrl ([WILDCARD])
+    at Worker.#pollControl ([WILDCARD])
