+++ cli/tests/workers/worker_nested_error.ts.out
@@ -2,4 +2,4 @@
     at foo ([WILDCARD])
     at [WILDCARD]
 error: Uncaught (in promise) Error: Unhandled error event reached main worker.
-    at Worker.#pollCtrl ([WILDCARD])
+    at Worker.#pollControl ([WILDCARD])
